fix!: change parameters and app db context for compatibility issues that made the program crash
using pluralizer and snake case to stick to good conventions

+++ MaceTech.API/Planning/Interfaces/REST/Resources/UpdateDevicePlantThresholdsResource.cs
@@ -5,8 +5,8 @@ namespace MaceTech.API.Planning.Interfaces.REST.Resources;
 
 public record UpdateDevicePlantThresholdsResource(
     [Required] Range<double> TemperaturaAmbiente,
-    [Required] Range<int> Humedad,
-    [Required] Range<int> Luminosidad,
+    [Required] Range<double> Humedad,
+    [Required] Range<double> Luminosidad,
     [Required] Range<double> SalinidadSuelo,
     [Required] Range<double> PhSuelo
 );
